+++ changelog.md
@@ -1,10 +1,38 @@
 
-## 1.1
+## 1.1 2020-06-27
 
-* 缩小了层级选择面板的按钮
+EN:
+
+* i18n supported (vue-i18n)
+
+* Optimized css selector of attribute "id"
+
+* Fixed wrong result of export "python lxml, css selector"
+
+* Supported attributes with newlines
+
+* Highlight with background-color
+
+* Layer limit panel: use smaller button
+
+* Element select panel & Layer limit panel: added scroller
+
+* Conclusion panel: css / xpath switcher
+
+中文:
+
+* i18n 支持 (vue-i18n)
 
 * 优化了 id 属性CSS选择器的生成
 
-* i18n 支持
+* 修正了导出的 python css lxml 语句不正确的问题
 
-* 层级选择和元素选择面板增加滚动条
+* 支持了属性中带换行符的元素
+
+* 高亮时带上背景色
+
+* 层级选择面板：缩小了按钮
+
+* 层级选择和元素选择面板：增加滚动条
+
+* 结论面板：css / xpath 切换选项卡